Refactored PackageService code, created AssetFrameworkService
